evitando usuário curtir as próprias sugestões

+++ services/aquisicao/aquisicaoapp/views/sugestao_aquisicao.py
@@ -18,27 +18,29 @@ class SugestaoAquisicaoViewSet(ModelViewSet):
 
     @action(methods=['put'], detail=True, url_path='curtida')
     def curtida(self, request, _id=None):
+        usuario_id = request.user['_id']
         sugestao = self.get_object()
 
-        with transaction.atomic():
-            aux = {
-                'sugestao_aquisicao_id': sugestao.pk,
-                'usuario_id': request.user['_id']
-            }
+        if str(sugestao.usuario_id) != usuario_id:
+            with transaction.atomic():
+                aux = {
+                    'sugestao_aquisicao_id': sugestao.pk,
+                    'usuario_id': usuario_id
+                }
 
-            c = Curtida.objects.filter(**aux).first()
-            if c is None:
-                SugestaoAquisicao.objects.filter(pk=sugestao.pk).update(
-                    quantidade_curtidas=models.F('quantidade_curtidas') + 1
-                )
-                aux['aviso'] = bool(request.GET.get('aviso'))
-                Curtida.objects.create(**aux)
+                c = Curtida.objects.filter(**aux).first()
+                if c is None:
+                    SugestaoAquisicao.objects.filter(pk=sugestao.pk).update(
+                        quantidade_curtidas=models.F('quantidade_curtidas') + 1
+                    )
+                    aux['aviso'] = bool(request.GET.get('aviso'))
+                    Curtida.objects.create(**aux)
 
-            else:
-                SugestaoAquisicao.objects.filter(pk=sugestao.pk).update(
-                    quantidade_curtidas=models.F('quantidade_curtidas') - 1
-                )
-                c.delete()
+                else:
+                    SugestaoAquisicao.objects.filter(pk=sugestao.pk).update(
+                        quantidade_curtidas=models.F('quantidade_curtidas') - 1
+                    )
+                    c.delete()
 
-            return Response(status=200)
+        return Response(status=200)
